progress on migration to types + constants files
