+++ components/avatar/types.ts
@@ -1,0 +1,7 @@
+import { ButtonHTMLAttributes } from 'react';
+
+export interface AvatarProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  children?: React.ReactNode;
+  className?: string;
+  bgImage?: string;
+}
