+++ components/button/constants.ts
@@ -1,0 +1,25 @@
+export const THEME = {
+  primary:
+    'text-black bg-blue-500 hover:bg-blue-400 active:bg-blue-300 border border-blue-500 hover:border-blue-400 active:border-blue-300',
+  'primary-alt':
+    'text-blue-500 bg-transparent hover:bg-transparent active:bg-transparent border border-blue-500 hover:border-blue-400 active:border-blue-300',
+
+  secondary:
+    'text-white bg-gray-500 hover:bg-gray-400 active:bg-gray-300 border border-gray-500 hover:border-gray-400 active:border-gray-300',
+  'secondary-alt':
+    'text-gray-300 bg-transparent hover:bg-transparent active:bg-transparent border border-gray-400 hover:border-gray-300 active:border-gray-200',
+
+  white:
+    'text-gray-700 bg-white hover:text-white hover:bg-transparent active:bg-transparent border border-gray-400 hover:border-gray-300 active:border-gray-200',
+
+  danger:
+    'text-red-700 bg-transparent hover:text-white hover:bg-red-700 active:bg-red-600 border border-red-700 hover:border-red-600 active:border-red-500',
+};
+
+export const SIZE = {
+  xs: 'text-sm px-2 py-0.5',
+  s: 'text-sm px-3 py-0.5',
+  base: 'text-sm px-8 py-2',
+  l: 'text-base px-8 py-3',
+  xl: 'text-base px-14 py-3',
+};
