+++ components/forms/label/constants.ts
@@ -1,0 +1,5 @@
+// eslint-disable-next-line import/prefer-default-export
+export const THEME = {
+  dark: 'block font-heading font-medium text-xs text-white',
+  light: 'block font-heading font-medium text-xs text-gray-600',
+};
