+++ components/icon/types.ts
@@ -1,0 +1,8 @@
+export interface IconProps {
+  icon: {
+    id: string;
+    viewBox: string;
+  };
+  className?: string;
+  style?: unknown;
+}
