+++ components/map/component.tsx
@@ -5,49 +5,14 @@ import cx from 'classnames';
 
 import isEmpty from 'lodash/isEmpty';
 
-import ReactMapGL, { FlyToInterpolator, TRANSITION_EVENTS, ViewportProps } from 'react-map-gl';
-import { InteractiveMapProps } from 'react-map-gl/src/components/interactive-map';
+import ReactMapGL, { FlyToInterpolator, TRANSITION_EVENTS } from 'react-map-gl';
 
 import { fitBounds } from '@math.gl/web-mercator';
 
 import { easeCubic } from 'd3-ease';
 
-export interface MapProps extends InteractiveMapProps {
-  /** A function that returns the map instance */
-  children?: React.ReactNode;
-
-  /** Custom css class for styling */
-  className?: string;
-
-  /** An object that defines the viewport
-   * @see https://uber.github.io/react-map-gl/#/Documentation/api-reference/interactive-map?section=initialization
-   */
-  viewport?: Partial<ViewportProps>;
-
-  /** An object that defines the bounds */
-  bounds?: {
-    bbox: number[];
-    options?: Record<string, unknown>;
-    viewportOptions?: Partial<ViewportProps>;
-  };
-
-  /** A function that exposes when the map is mounted.
-   * It receives and object with the `mapRef` and `mapContainerRef` reference. */
-  onMapReady?: ({ map, mapContainer }) => void;
-
-  /** A function that exposes when the map is loaded.
-   * It receives and object with the `mapRef` and `mapContainerRef` reference. */
-  onMapLoad?: ({ map, mapContainer }) => void;
-
-  /** A function that exposes the viewport */
-  onMapViewportChange?: (viewport: Partial<ViewportProps>) => void;
-}
-
-const DEFAULT_VIEWPORT = {
-  zoom: 2,
-  latitude: 0,
-  longitude: 0,
-};
+import { MapProps } from './types';
+import { DEFAULT_VIEWPORT } from './constants';
 
 export const Map: FC<MapProps> = ({
   mapboxApiAccessToken,
