+++ components/map/constants.ts
@@ -1,0 +1,6 @@
+// eslint-disable-next-line import/prefer-default-export
+export const DEFAULT_VIEWPORT = {
+  zoom: 2,
+  latitude: 0,
+  longitude: 0,
+};
