+++ components/map/controls/fit-bounds/component.tsx
@@ -3,19 +3,9 @@ import cx from 'classnames';
 
 import Icon from 'components/icon';
 
-import { ViewportProps } from 'react-map-gl';
-
 import FIT_BOUNDS_SVG from 'svgs/map/fit-bounds.svg?sprite';
 
-export interface FitBoundsControlProps {
-  bounds?: {
-    bbox?: number[];
-    options?: Record<string, unknown>;
-    viewportOptions?: Partial<ViewportProps>;
-  };
-  className?: string;
-  onFitBoundsChange: (bounds) => void;
-}
+import { FitBoundsControlProps } from './types';
 
 export const FitBoundsControl: FC<FitBoundsControlProps> = ({
   bounds,
