+++ components/map/controls/fit-bounds/types.ts
@@ -1,0 +1,11 @@
+import { ViewportProps } from 'react-map-gl';
+
+export interface FitBoundsControlProps {
+  bounds?: {
+    bbox?: number[];
+    options?: Record<string, unknown>;
+    viewportOptions?: Partial<ViewportProps>;
+  };
+  className?: string;
+  onFitBoundsChange: (bounds) => void;
+}
