+++ components/map/controls/types.ts
@@ -1,0 +1,4 @@
+export interface ControlsProps {
+  className?: string;
+  children: React.ReactNode;
+}
