+++ components/map/controls/zoom/types.ts
@@ -1,0 +1,7 @@
+import { ViewportProps } from 'react-map-gl';
+
+export interface ZoomControlProps {
+  viewport: Partial<ViewportProps>;
+  className?: string;
+  onZoomChange: (zoom: number) => void;
+}
