+++ components/search/component.tsx
@@ -6,27 +6,12 @@ import { useSearchField } from '@react-aria/searchfield';
 import { useSearchFieldState } from '@react-stately/searchfield';
 import { useButton } from '@react-aria/button'; // eslint-disable-line @typescript-eslint/no-unused-vars
 
-// react types
-import { AriaSearchFieldProps } from '@react-types/searchfield';
-
 import Icon from 'components/icon';
 import SEARCH_SVG from 'svgs/ui/search.svg?sprite';
 import CLOSE_SVG from 'svgs/ui/close.svg?sprite'; // eslint-disable-line @typescript-eslint/no-unused-vars
 
-const THEME = {
-  dark: 'text-white',
-  light: 'text-black',
-};
-
-const SIZES = {
-  sm: 'text-sm h-10',
-  base: 'text-base h-12',
-};
-
-export interface SearchProps extends AriaSearchFieldProps {
-  theme?: 'dark' | 'light';
-  size: 'sm' | 'base';
-}
+import { SearchProps } from './types';
+import { SIZES, THEME } from './constants';
 
 export const Search: FC<SearchProps> = ({
   theme = 'dark',
