+++ components/search/constants.ts
@@ -1,0 +1,9 @@
+export const THEME = {
+  dark: 'text-white',
+  light: 'text-black',
+};
+
+export const SIZES = {
+  sm: 'text-sm h-10',
+  base: 'text-base h-12',
+};
