+++ components/tag/types.ts
@@ -1,0 +1,4 @@
+export interface TagProps {
+  children: React.ReactNode;
+  className?: string | unknown;
+}
